separate fpeutils
