+++ fpeutils/numeral_test.go
@@ -16,7 +16,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and limitations under the License.
 
 */
-package fpe
+
+// Package fpeutils provides some encoding helpers for use
+// in the FF1 and FF3 format-preserving encryption packages.
+package fpeutils
 
 import (
 	"fmt"
